move to env config

+++ app/routers/post.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI, Response, status, HTTPException, Depends, APIRouter
-from typing import List
+from typing import List, Optional
 from sqlalchemy.orm import Session
 import logging
 import schemas
@@ -19,13 +19,13 @@ router = APIRouter(
 
 # Decorator - turns function into a path operator
 @router.get("/", response_model=List[schemas.Post])
-def get_posts(db: Session = Depends(database.get_db), current_user: models.User = Depends(oauth2.get_current_user)):
-    posts = db.query(models.Post).all()
+def get_posts(db: Session = Depends(database.get_db), current_user: models.User = Depends(oauth2.get_current_user), limit: int = 10, skip: int = 0, search: Optional[str] = ""):
+    posts = db.query(models.Post).filter(models.Post.title.contains(search)).limit(limit).offset(skip).all()
     return posts
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
 def create_posts(post: schemas.PostCreate, db: Session = Depends(database.get_db), current_user: models.User = Depends(oauth2.get_current_user)):
-    model_post = models.Post(**post.model_dump())
+    model_post = models.Post(owner_id=current_user.id, **post.model_dump())
     db.add(model_post)
     db.commit()
     db.refresh(model_post)
@@ -49,6 +49,9 @@ def update_post(id: int, post: schemas.PostCreate, db: Session = Depends(databas
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} does not exist")
     
+    if post_query.first().owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
     
     post_query.update(post.model_dump(), synchronize_session=False)
     db.commit()
@@ -62,6 +65,10 @@ def delete_post(id: int, db: Session = Depends(database.get_db), current_user: m
     if post_query.first() == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} was not found")
+    
+    if post_query.first().owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requested action")
     
     post_query.delete(synchronize_session=False)
     db.commit()
